push/requirements.txt
APScheduler==3.11.0
python-dotenv==1.2.1
python-ntfy==0.10.1
requests==2.32.5
pydantic